@@ -12,10 +12,10 @@
     "@svgr/webpack": "^6.3.1",
     "@vercel/analytics": "^0.1.11",
     "neo4j-driver": "^5.7.0",
-    "next": "12.2.2",
+    "next": "^13.4.7",
     "nextjs-google-analytics": "^2.1.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "satisfactory-docs-parser": "^6.0.0",
     "styled-components": "^5.3.5"
   },
@@ -23,9 +23,12 @@
     "@types/node": "^18.11.18",
     "@types/react": "^18.0.26",
     "@types/styled-components": "^5.1.26",
+    "autoprefixer": "^10.4.14",
     "babel-plugin-styled-components": "^2.0.7",
     "eslint": "8.19.0",
-    "eslint-config-next": "12.2.2",
+    "eslint-config-next": "^13.4.7",
+    "postcss": "^8.4.24",
+    "tailwindcss": "^3.3.2",
     "typescript": "^4.9.4"
   }
 }
